allow passing "weak arguments" to signal callbacks
This means that the arguments are passed just like the existing
user_args, but they are stored as weak references to prevent a callback
from keeping its user_args alive. The emit code transparently
dereferences the arguments before passing them to the callback.

If any of the arguments are cleaned up, the callback is no longer called
(it's not actually automatically removed from the list yet, but it will
be).