Reset screen to original state on any exception.
In MainLoop._run() we were running self.event_loop.run() in
an exception frame so that we could reset the screen in the
case of an unclean shutdown of the MainLoop. The catch specification
was "except Exception as e" but this doesn't include exceptions that
don't derive from Exception, like KeyboardInterrupt (i.e. Ctrl-C).
This change fixes that.