Add 256-color support for fbterm
fbterm supports an analog of xterm's 256 color method, but due to
overlap in the control codes, it uses different escape sequences.
Since urwid doesn't use terminfo, it has to be coded in the raw_display
directly.

The call to register_palette_entry() during __init__ was moved because
it now relies on the self.term attribute, and was previously being
invoked before the class was fully initialized.