Do not emit change until after text is updated
Sometimes an Edit's change handler wants to edit the text that has just
been entered.  For instance, if the handler does validation, it may want
to immediately highlight or prevent invalid entries from being shown.
This is not possible if the change signal is emitted before the internal
text data is updated as the update will overwrite anything that the
change handler performed.  Emitting the change event after the update
has been performed allows for changes to be made.

Fixes #212