Implement postchange signal for Edit widget.
The Edit widget in 1.3.0 has a "change" signal which fires before the
edit_text is updated.  This means that change signal handlers cannot
update the edit_text; the text is updated by the signal emitter after
the change handler has finished running.  To remedy this, in a backwards
compatible way, create a second signal, "postchange" which fires after
edit_text has been changed.  "postchange" sends the widget's old text
value as a parameter so that change handlers can use the former value
if needed (for instance, to revert a change because the new_text was
invalid)