checkbox and radiobutton also needs to have a postchange event
Just like Edit() boxes, CheckBox and RadioButton need to have an event
emitted after the change has been commited.  This allows us to modify
the state after it occurs.  (If the event only fires before the state is
actually changed, the event handler cannot change the state because the
emitter will change it to what they were going to after the event
handler has finished.