Fix #246, fix #234
Adjusts the internal representation of an alarm in SelectEventLoop to
include a "tie-break" value prior to the callback (as recommended in the
heapq documentation) to prevent TypeError when comparing two different
callbacks. Also tweaks SelectEventLoop to use heapq.heappop instead of
list.pop for retrieving the next alarm, to maintain the heap structure
of the alarm list.