Fix(docs): screenshot rendering race-condition
Fixes #101 and #102.

The issue is rather simple:
 * we run an example in a virtual X display,
 * we send events to an example via xdotool,
 * immediately after, we take a screenshot (with ImageMagick).

We don't ever synchronize to a point "all X11 events have
been pumped, handled, and the example has redrawn" -- so there's an
obvious race happening.

Fix the race by locating the PID of the example interpreter process,
(it'll be the only child of our urxvt instance),
and polling its process status (running or not). When we detect it's
not running N times in a row -- conclude it's done and went idling.
With increasing N, the total probability of misdetection drops
exponentially.

N=10 should suffice as a rough 1st guess; if the race reappears --
simply increase it a couple of points, say to 15 or 16.