Find nearest approximation for 24-bit colors in 256-color AttrSpecs
Palette entries can only distinguish between "low" and "high" colors, so
there's no way for a user to specify different colors for 88, 256, and
24-bit color screens.  Prior to this change, someone wanting to specify
a 24-bit color but have their applicaiton also work when running in
256-color mode would have to maintain two different palette entries, or
do a color conversion themselves to avoid an AttrSpec error.

With this change, an AttrSpec with colors=256 will instead convert
24-bit foreground and/or background colors to the nearest color from the
256-color palette.  In most cases this is what the user will want, and
if they want to handle 256-color mode separately, they can still do so.
This just makes it easier to support both 24-bit and 8-bit color modes.