Add bin/release.sh script to partially automate releases.
This script takes the versino number to release as a single argument.
It tries to generate a sensible changelog from commit messages, but
allows for manual cleanup.  It also uses the `contributors` Python
module to generate a list of contributors to be included in the README.

Error checking is present but probably insufficient, but at the very
least, this should serve as a document of the current release process.