Fix handling of WEIGHT selectable items in the Pile
In the case of not enough rows for full render,
focused widget should receive priority
and need to be rendered in the price of dropping other widgets.

For the optimization purposes and better code organization,
protected `_get_flow_rows_sizes` was detached from the `get_rows_sizes`.
It should not be called manually.

Fix #1005